Dataframe script and csv
multi_lang_article_pandas.csv should be the dataframe containing all articles in all languages requested. Any articles not containing a link to a specific language is filled with "NONE" by default.  This is important to note for next step of processing. 

multi_lang_article_pandas.py is the script to get raw text links (in raw_article_links.txt) to a pandas dataframe and to a csv.